refactor(lib): extract lib/auth_password.php — password + reset tokens (ADR-004, sub of #907)
Moves validate_password_complexity and the password-reset token/email helpers
out of lib.php into lib/auth_password.php. ADR-003 conversion applied.
ipam_argon2id_derive stays in lib.php — it is the backup-vault KDF, not a
password function (the plan's Task 6.2 anchor list mis-categorized it).
Rate-limit and reCAPTCHA functions stay behind for Tasks 6.3-6.4.
